Add files via upload
This script is used to shift specific columns of data a specific number of seconds. 
The purpose of this is to be able to use an easily identifiable event trigger to line up data generated from different test runs. 
This code uses the file 3_event_triggers.txt to show basic functionality.